Revert "[in_app_purchase] migrate to the v2 android embedding (#2155)"
This reverts commit c794ac18405e0dfa631dfe5c5b8f076a861e07c3.
