+++ packages/in_app_purchase/CHANGELOG.md
@@ -1,10 +1,3 @@
-## 0.2.2
-
-* Support the v2 Android embedder.
-* Update to AndroidX.
-* Migrate to using the new e2e test binding.
-* Add a e2e test.
-
 ## 0.2.1+5
 
 * Define clang module for iOS.
