+++ packages/in_app_purchase/android/src/main/java/io/flutter/plugins/inapppurchase/BillingClientFactory.java
@@ -1,23 +1,19 @@
-// Copyright 2019 The Chromium Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
 package io.flutter.plugins.inapppurchase;
 
 import android.content.Context;
-import androidx.annotation.NonNull;
 import com.android.billingclient.api.BillingClient;
 import io.flutter.plugin.common.MethodChannel;
 
-/** Responsible for creating a {@link BillingClient} object. */
 interface BillingClientFactory {
+  BillingClient createBillingClient(Context context, MethodChannel channel);
+}
 
-  /**
-   * Creates and returns a {@link BillingClient}.
-   *
-   * @param context The context used to create the {@link BillingClient}.
-   * @param channel The method channel used to create the {@link BillingClient}.
-   * @return The {@link BillingClient} object that is created.
-   */
-  BillingClient createBillingClient(@NonNull Context context, @NonNull MethodChannel channel);
+final class BillingClientFactoryImpl implements BillingClientFactory {
+
+  @Override
+  public BillingClient createBillingClient(Context context, MethodChannel channel) {
+    return BillingClient.newBuilder(context)
+        .setListener(new PluginPurchaseListener(channel))
+        .build();
+  }
 }
