+++ packages/in_app_purchase/example/android/app/src/main/java/com/example/inapppurchaseexample/MainActivity.java
@@ -8,7 +8,7 @@ import android.os.Bundle;
 import io.flutter.app.FlutterActivity;
 import io.flutter.plugins.GeneratedPluginRegistrant;
 
-public class EmbeddingV1Activity extends FlutterActivity {
+public class MainActivity extends FlutterActivity {
   @Override
   protected void onCreate(Bundle savedInstanceState) {
     super.onCreate(savedInstanceState);
