+++ packages/in_app_purchase/example/android/app/src/test/java/io/flutter/plugins/inapppurchase/InAppPurchasePluginTest.java
@@ -29,7 +29,6 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import android.app.Activity;
-import android.content.Context;
 import androidx.annotation.Nullable;
 import com.android.billingclient.api.BillingClient;
 import com.android.billingclient.api.BillingClient.BillingResponse;
@@ -46,6 +45,7 @@ import com.android.billingclient.api.SkuDetailsResponseListener;
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
 import io.flutter.plugin.common.MethodChannel.Result;
+import io.flutter.plugin.common.PluginRegistry;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -56,27 +56,26 @@ import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 import org.mockito.Spy;
 
-public class MethodCallHandlerTest {
-  private MethodCallHandlerImpl methodChannelHandler;
-  private BillingClientFactory factory;
+public class InAppPurchasePluginTest {
+  private InAppPurchasePlugin plugin;
   @Mock BillingClient mockBillingClient;
   @Mock MethodChannel mockMethodChannel;
   @Spy Result result;
+  @Mock PluginRegistry.Registrar registrar;
   @Mock Activity activity;
-  @Mock Context context;
 
   @Before
   public void setUp() {
     MockitoAnnotations.initMocks(this);
 
-    factory = (context, channel) -> mockBillingClient;
-    methodChannelHandler = new MethodCallHandlerImpl(activity, context, mockMethodChannel, factory);
+    BillingClientFactory factory = (context, channel) -> mockBillingClient;
+    plugin = new InAppPurchasePlugin(factory, registrar, mockMethodChannel);
   }
 
   @Test
   public void invalidMethod() {
     MethodCall call = new MethodCall("invalid", null);
-    methodChannelHandler.onMethodCall(call, result);
+    plugin.onMethodCall(call, result);
     verify(result, times(1)).notImplemented();
   }
 
@@ -85,7 +84,7 @@ public class MethodCallHandlerTest {
     mockStartConnection();
     MethodCall call = new MethodCall(IS_READY, null);
     when(mockBillingClient.isReady()).thenReturn(true);
-    methodChannelHandler.onMethodCall(call, result);
+    plugin.onMethodCall(call, result);
     verify(result).success(true);
   }
 
@@ -94,17 +93,17 @@ public class MethodCallHandlerTest {
     mockStartConnection();
     MethodCall call = new MethodCall(IS_READY, null);
     when(mockBillingClient.isReady()).thenReturn(false);
-    methodChannelHandler.onMethodCall(call, result);
+    plugin.onMethodCall(call, result);
     verify(result).success(false);
   }
 
   @Test
   public void isReady_clientDisconnected() {
     MethodCall disconnectCall = new MethodCall(END_CONNECTION, null);
-    methodChannelHandler.onMethodCall(disconnectCall, mock(Result.class));
+    plugin.onMethodCall(disconnectCall, mock(Result.class));
     MethodCall isReadyCall = new MethodCall(IS_READY, null);
 
-    methodChannelHandler.onMethodCall(isReadyCall, result);
+    plugin.onMethodCall(isReadyCall, result);
 
     verify(result).error(contains("UNAVAILABLE"), contains("BillingClient"), any());
     verify(result, never()).success(any());
@@ -128,7 +127,7 @@ public class MethodCallHandlerTest {
         ArgumentCaptor.forClass(BillingClientStateListener.class);
     doNothing().when(mockBillingClient).startConnection(captor.capture());
 
-    methodChannelHandler.onMethodCall(call, result);
+    plugin.onMethodCall(call, result);
     verify(result, never()).success(any());
     captor.getValue().onBillingSetupFinished(100);
     captor.getValue().onBillingSetupFinished(200);
@@ -148,12 +147,12 @@ public class MethodCallHandlerTest {
     ArgumentCaptor<BillingClientStateListener> captor =
         ArgumentCaptor.forClass(BillingClientStateListener.class);
     doNothing().when(mockBillingClient).startConnection(captor.capture());
-    methodChannelHandler.onMethodCall(connectCall, mock(Result.class));
+    plugin.onMethodCall(connectCall, mock(Result.class));
     final BillingClientStateListener stateListener = captor.getValue();
 
     // Disconnect the connected client
     MethodCall disconnectCall = new MethodCall(END_CONNECTION, null);
-    methodChannelHandler.onMethodCall(disconnectCall, result);
+    plugin.onMethodCall(disconnectCall, result);
 
     // Verify that the client is disconnected and that the OnDisconnect callback has
     // been triggered
@@ -177,7 +176,7 @@ public class MethodCallHandlerTest {
     MethodCall queryCall = new MethodCall(QUERY_SKU_DETAILS, arguments);
 
     // Query for SKU details
-    methodChannelHandler.onMethodCall(queryCall, result);
+    plugin.onMethodCall(queryCall, result);
 
     // Assert the arguments were forwarded correctly to BillingClient
     ArgumentCaptor<SkuDetailsParams> paramCaptor = ArgumentCaptor.forClass(SkuDetailsParams.class);
@@ -202,7 +201,7 @@ public class MethodCallHandlerTest {
   public void querySkuDetailsAsync_clientDisconnected() {
     // Disconnect the Billing client and prepare a querySkuDetails call
     MethodCall disconnectCall = new MethodCall(END_CONNECTION, null);
-    methodChannelHandler.onMethodCall(disconnectCall, mock(Result.class));
+    plugin.onMethodCall(disconnectCall, mock(Result.class));
     String skuType = BillingClient.SkuType.INAPP;
     List<String> skusList = asList("id1", "id2");
     HashMap<String, Object> arguments = new HashMap<>();
@@ -211,7 +210,7 @@ public class MethodCallHandlerTest {
     MethodCall queryCall = new MethodCall(QUERY_SKU_DETAILS, arguments);
 
     // Query for SKU details
-    methodChannelHandler.onMethodCall(queryCall, result);
+    plugin.onMethodCall(queryCall, result);
 
     // Assert that we sent an error back.
     verify(result).error(contains("UNAVAILABLE"), contains("BillingClient"), any());
@@ -220,6 +219,7 @@ public class MethodCallHandlerTest {
 
   @Test
   public void launchBillingFlow_ok_nullAccountId() {
+    when(registrar.activity()).thenReturn(activity);
     // Fetch the sku details first and then prepare the launch billing flow call
     String skuId = "foo";
     queryForSkus(singletonList(skuId));
@@ -231,7 +231,7 @@ public class MethodCallHandlerTest {
     // Launch the billing flow
     int responseCode = BillingResponse.OK;
     when(mockBillingClient.launchBillingFlow(any(), any())).thenReturn(responseCode);
-    methodChannelHandler.onMethodCall(launchCall, result);
+    plugin.onMethodCall(launchCall, result);
 
     // Verify we pass the arguments to the billing flow
     ArgumentCaptor<BillingFlowParams> billingFlowParamsCaptor =
@@ -248,8 +248,6 @@ public class MethodCallHandlerTest {
 
   @Test
   public void launchBillingFlow_ok_null_Activity() {
-    methodChannelHandler.setActivity(null);
-
     // Fetch the sku details first and then prepare the launch billing flow call
     String skuId = "foo";
     String accountId = "account";
@@ -258,7 +256,8 @@ public class MethodCallHandlerTest {
     arguments.put("sku", skuId);
     arguments.put("accountId", accountId);
     MethodCall launchCall = new MethodCall(LAUNCH_BILLING_FLOW, arguments);
-    methodChannelHandler.onMethodCall(launchCall, result);
+
+    plugin.onMethodCall(launchCall, result);
 
     // Verify we pass the response code to result
     verify(result).error(contains("ACTIVITY_UNAVAILABLE"), contains("foreground"), any());
@@ -267,6 +266,7 @@ public class MethodCallHandlerTest {
 
   @Test
   public void launchBillingFlow_ok_AccountId() {
+    when(registrar.activity()).thenReturn(activity);
     // Fetch the sku details first and query the method call
     String skuId = "foo";
     String accountId = "account";
@@ -279,7 +279,7 @@ public class MethodCallHandlerTest {
     // Launch the billing flow
     int responseCode = BillingResponse.OK;
     when(mockBillingClient.launchBillingFlow(any(), any())).thenReturn(responseCode);
-    methodChannelHandler.onMethodCall(launchCall, result);
+    plugin.onMethodCall(launchCall, result);
 
     // Verify we pass the arguments to the billing flow
     ArgumentCaptor<BillingFlowParams> billingFlowParamsCaptor =
@@ -298,7 +298,7 @@ public class MethodCallHandlerTest {
   public void launchBillingFlow_clientDisconnected() {
     // Prepare the launch call after disconnecting the client
     MethodCall disconnectCall = new MethodCall(END_CONNECTION, null);
-    methodChannelHandler.onMethodCall(disconnectCall, mock(Result.class));
+    plugin.onMethodCall(disconnectCall, mock(Result.class));
     String skuId = "foo";
     String accountId = "account";
     HashMap<String, Object> arguments = new HashMap<>();
@@ -306,7 +306,7 @@ public class MethodCallHandlerTest {
     arguments.put("accountId", accountId);
     MethodCall launchCall = new MethodCall(LAUNCH_BILLING_FLOW, arguments);
 
-    methodChannelHandler.onMethodCall(launchCall, result);
+    plugin.onMethodCall(launchCall, result);
 
     // Assert that we sent an error back.
     verify(result).error(contains("UNAVAILABLE"), contains("BillingClient"), any());
@@ -324,7 +324,7 @@ public class MethodCallHandlerTest {
     arguments.put("accountId", accountId);
     MethodCall launchCall = new MethodCall(LAUNCH_BILLING_FLOW, arguments);
 
-    methodChannelHandler.onMethodCall(launchCall, result);
+    plugin.onMethodCall(launchCall, result);
 
     // Assert that we sent an error back.
     verify(result).error(contains("NOT_FOUND"), contains(skuId), any());
@@ -342,7 +342,7 @@ public class MethodCallHandlerTest {
 
     HashMap<String, Object> arguments = new HashMap<>();
     arguments.put("skuType", SkuType.INAPP);
-    methodChannelHandler.onMethodCall(new MethodCall(QUERY_PURCHASES, arguments), result);
+    plugin.onMethodCall(new MethodCall(QUERY_PURCHASES, arguments), result);
 
     // Verify we pass the response to result
     ArgumentCaptor<HashMap<String, Object>> resultCaptor = ArgumentCaptor.forClass(HashMap.class);
@@ -354,11 +354,11 @@ public class MethodCallHandlerTest {
   @Test
   public void queryPurchases_clientDisconnected() {
     // Prepare the launch call after disconnecting the client
-    methodChannelHandler.onMethodCall(new MethodCall(END_CONNECTION, null), mock(Result.class));
+    plugin.onMethodCall(new MethodCall(END_CONNECTION, null), mock(Result.class));
 
     HashMap<String, Object> arguments = new HashMap<>();
     arguments.put("skuType", SkuType.INAPP);
-    methodChannelHandler.onMethodCall(new MethodCall(QUERY_PURCHASES, arguments), result);
+    plugin.onMethodCall(new MethodCall(QUERY_PURCHASES, arguments), result);
 
     // Assert that we sent an error back.
     verify(result).error(contains("UNAVAILABLE"), contains("BillingClient"), any());
@@ -377,8 +377,7 @@ public class MethodCallHandlerTest {
     ArgumentCaptor<PurchaseHistoryResponseListener> listenerCaptor =
         ArgumentCaptor.forClass(PurchaseHistoryResponseListener.class);
 
-    methodChannelHandler.onMethodCall(
-        new MethodCall(QUERY_PURCHASE_HISTORY_ASYNC, arguments), result);
+    plugin.onMethodCall(new MethodCall(QUERY_PURCHASE_HISTORY_ASYNC, arguments), result);
 
     // Verify we pass the data to result
     verify(mockBillingClient)
@@ -393,12 +392,11 @@ public class MethodCallHandlerTest {
   @Test
   public void queryPurchaseHistoryAsync_clientDisconnected() {
     // Prepare the launch call after disconnecting the client
-    methodChannelHandler.onMethodCall(new MethodCall(END_CONNECTION, null), mock(Result.class));
+    plugin.onMethodCall(new MethodCall(END_CONNECTION, null), mock(Result.class));
 
     HashMap<String, Object> arguments = new HashMap<>();
     arguments.put("skuType", SkuType.INAPP);
-    methodChannelHandler.onMethodCall(
-        new MethodCall(QUERY_PURCHASE_HISTORY_ASYNC, arguments), result);
+    plugin.onMethodCall(new MethodCall(QUERY_PURCHASE_HISTORY_ASYNC, arguments), result);
 
     // Assert that we sent an error back.
     verify(result).error(contains("UNAVAILABLE"), contains("BillingClient"), any());
@@ -432,7 +430,7 @@ public class MethodCallHandlerTest {
     ArgumentCaptor<ConsumeResponseListener> listenerCaptor =
         ArgumentCaptor.forClass(ConsumeResponseListener.class);
 
-    methodChannelHandler.onMethodCall(new MethodCall(CONSUME_PURCHASE_ASYNC, arguments), result);
+    plugin.onMethodCall(new MethodCall(CONSUME_PURCHASE_ASYNC, arguments), result);
 
     // Verify we pass the data to result
     verify(mockBillingClient).consumeAsync(eq("mockToken"), listenerCaptor.capture());
@@ -453,7 +451,7 @@ public class MethodCallHandlerTest {
         ArgumentCaptor.forClass(BillingClientStateListener.class);
     doNothing().when(mockBillingClient).startConnection(captor.capture());
 
-    methodChannelHandler.onMethodCall(call, result);
+    plugin.onMethodCall(call, result);
     return captor;
   }
 
@@ -468,7 +466,7 @@ public class MethodCallHandlerTest {
     }
 
     MethodCall connectCall = new MethodCall(START_CONNECTION, arguments);
-    methodChannelHandler.onMethodCall(connectCall, result);
+    plugin.onMethodCall(connectCall, result);
   }
 
   private void queryForSkus(List<String> skusList) {
@@ -481,7 +479,7 @@ public class MethodCallHandlerTest {
     MethodCall queryCall = new MethodCall(QUERY_SKU_DETAILS, arguments);
 
     // Call the method.
-    methodChannelHandler.onMethodCall(queryCall, mock(Result.class));
+    plugin.onMethodCall(queryCall, mock(Result.class));
 
     // Respond to the call with a matching set of Sku details.
     ArgumentCaptor<SkuDetailsResponseListener> listenerCaptor =
