+++ packages/in_app_purchase/example/android/gradle.properties
@@ -1,4 +1,1 @@
 org.gradle.jvmargs=-Xmx1536M
-android.enableR8=true
-android.useAndroidX=true
-android.enableJetifier=true
