+++ packages/in_app_purchase/pubspec.yaml
@@ -2,7 +2,7 @@ name: in_app_purchase
 description: A Flutter plugin for in-app purchases. Exposes APIs for making in-app purchases through the App Store and Google Play.
 author:  Flutter Team <flutter-dev@googlegroups.com>
 homepage: https://github.com/flutter/plugins/tree/master/packages/in_app_purchase
-version: 0.2.2
+version: 0.2.1+5
 
 
 dependencies:
@@ -24,7 +24,6 @@ dev_dependencies:
     path: example/
   test: ^1.5.2
   shared_preferences: ^0.5.2
-  e2e: ^0.2.0
 
 flutter:
   plugin:
@@ -33,4 +32,4 @@ flutter:
 
 environment:
   sdk: ">=2.0.0-dev.28.0 <3.0.0"
-  flutter: ">=1.6.7 <2.0.0"
+  flutter: ">=1.5.0 <2.0.0"
